+++ labs/celery.py
@@ -9,13 +9,8 @@ import json
 
 from labs.database.embeddings import find_similar_embeddings
 from labs.database.vectorize import vectorize_to_database
-from labs.repo import clone_repository
-from labs.middleware import (
-    call_agent_to_apply_code_changes,
-    get_llm_response,
-    get_prompt,
-    prepare_context,
-)
+from labs.repo import call_agent_to_apply_code_changes, clone_repository
+from labs.middleware import get_llm_response, get_prompt, prepare_context
 from labs.run import commit_changes, create_branch, create_pull_request, get_issue
