+++ labs/run.py
@@ -1,10 +1,16 @@
 import logging
 
-from labs.repo import clone_repository
 from labs.decorators import time_and_log_function
-from labs.middleware import call_llm_with_context, call_agent_to_apply_code_changes
+from labs.llm import call_llm_with_context
 from labs.config import settings
-from labs.repo import commit_changes, create_branch, create_pull_request, get_issue
+from labs.repo import (
+    commit_changes,
+    create_branch,
+    create_pull_request,
+    get_issue,
+    call_agent_to_apply_code_changes,
+    clone_repository,
+)
 
 
 logger = logging.getLogger(__name__)
